+++ client/src/protocol.js
@@ -1,44 +1,44 @@
 /**
- * Replicated from server! See server/src/protocol.js. 
+ * Replicated from server. See ../server/protocol.js for original.
  * 
  * Protocol for websocket communications between server and client.
  * All requests and responses are JSON-encoded strings.
  * 
  * Required fields for client actions: 
  * HOST: name: string
- * JOIN: name: string, id: string
+ * JOIN: id: string, name: string
  * RESPOND: selection: int
  * 
  * Recieved field from server actions: 
- * HOSTED: name: string, id: string, players: [string]
- * JOINED: name: string, id: string, players: [string]
- * LEFT: name: string, id: string, players: [string]
+ * HOSTED: id: string, name: string, players: [string]
+ * JOINED: id: string, name: string, players: [string]
+ * LEFT: id: string, name: string, players: [string]
  * RESPONDED: name: string
  * QUESTION: choices: [string]
  * ANSWER: answer: int, correct: bool, leaderboard: [(string, int)]
  * ENDED: leaderboard: [(string, int)]
  * ERROR: message: string
-
  */
 export const Actions = Object.freeze({
-  NONE: 'none',
+  NONE: 'NONE',
   /* client to server actions */
-  HOST: 'host',
-  JOIN: 'join',
-  LEAVE: 'leave',
-  START: 'start',
-  CANCEL: 'cancel',
-  RESPOND: 'respond',
-  READY: 'ready',
+  HOST: 'HOST',
+  JOIN: 'JOIN',
+  LEAVE: 'LEAVE',
+  START: 'START',
+  CANCEL: 'CANCEL',
+  RESPOND: 'RESPOND',
+  READY: 'READY',
   /* server to client actions */
-  HOSTED: 'hosted',
-  JOINED: 'joined',
-  LEFT: 'left',
-  STARTED: 'started',
-  CANCELLED: 'cancelled',
-  RESPONDED: 'responded',
-  QUESTION: 'question',
-  ANSWER: 'answer',
-  ENDED: 'ended',
-  ERROR: 'error',
+  HOSTED: 'HOSTED',
+  JOINED: 'JOINED',
+  LEFT: 'LEFT',
+  STARTED: 'STARTED',
+  CANCELLED: 'CANCELLED',
+  RESPONDED: 'RESPONDED',
+  QUESTION: 'QUESTION',
+  ANSWER: 'ANSWER',
+  ENDED: 'ENDED',
+  ERROR: 'ERROR',
 }); 
+